代码优化，删除多余的case type

+++ passport-main-common/src/main/java/com/sogou/upd/passport/common/utils/DateUtil.java
@@ -224,7 +224,7 @@ public class DateUtil {
         }
 
         long intervalMilli = lastDate.getTime() - firstDate.getTime();
-        return (long) (intervalMilli / (1000));
+        return (intervalMilli / (1000));
     }
 
     /**
